test: deduplicate net perf test code.
